修复表盘传输异常问题，2.2.70.15；
1.添加定制需求相关接口；
2.修复表盘传输异常问题；

+++ iOS_sdk_source/Framework/2.2.XX.15/VeepooBleSDK.framework/Headers/VPQH15GNSSLocationModel.h
@@ -14,7 +14,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface VPQH15GNSSLocationModel : NSObject
 
 //蓝牙RSSI
-@property (nonatomic, assign) uint8_t rssi;
+@property (nonatomic, assign) int rssi;
 
 //经纬度
 @property (nonatomic, assign) CLLocationCoordinate2D location;
